requirements.txt
py4j==0.10.9.7
pyspark==3.5.5
delta-spark==3.3.1
kafka-python==2.2.7
confluent-kafka==2.10.0
python-dotenv==1.1.0